requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-yasg==1.21.11
inflection==0.5.1
django-filter==24.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
openpyxl==3.1.5
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.4
uritemplate==4.2.0
